requirements.txt
# ── Core audio I/O ───────────────────────────────────────────
sounddevice==0.4.6
scipy==1.13.0
numpy==1.26.4
pydub==0.25.1
pygame==2.5.2

# ── ASR (Speech-to-Text) ─────────────────────────────────────
faster-whisper==1.0.3
openai-whisper==20231117

# ── LLM (Groq + OpenAI fallback support) ─────────────────────
groq==0.9.0
openai==1.30.1

# ── TTS (Text-to-Speech) ─────────────────────────────────────
gTTS==2.5.1
pyttsx3==2.90

# ── Environment config ───────────────────────────────────────
python-dotenv==1.0.1

# ── Networking / SSL stability ───────────────────────────────
httpx
certifi
huggingface_hub

# ── Optional but useful for audio processing ────────────────
# ffmpeg must be installed separately on system